neon_racer/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
cloudpickle==3.1.2
    # via
    #   gymnasium
    #   stable-baselines3
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
farama-notifications==0.0.4
    # via gymnasium
filelock==3.20.1
    # via torch
fonttools==4.61.1
    # via matplotlib
fsspec==2025.12.0
    # via torch
gymnasium==0.29.1
    # via
    #   -r requirements.in
    #   stable-baselines3
jinja2==3.1.6
    # via torch
kiwisolver==1.4.9
    # via matplotlib
markupsafe==3.0.3
    # via jinja2
matplotlib==3.10.8
    # via
    #   -r requirements.in
    #   stable-baselines3
mpmath==1.3.0
    # via sympy
neat-python==0.92
    # via -r requirements.in
networkx==3.6.1
    # via torch
numpy==1.26.4
    # via
    #   -r requirements.in
    #   contourpy
    #   gymnasium
    #   matplotlib
    #   pandas
    #   stable-baselines3
packaging==25.0
    # via matplotlib
pandas==2.3.3
    # via stable-baselines3
pillow==12.0.0
    # via matplotlib
pygame==2.6.1
    # via -r requirements.in
pyparsing==3.2.5
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
six==1.17.0
    # via python-dateutil
stable-baselines3==2.3.2
    # via -r requirements.in
sympy==1.14.0
    # via torch
torch==2.9.1
    # via stable-baselines3
typing-extensions==4.15.0
    # via
    #   gymnasium
    #   torch
tzdata==2025.3
    # via pandas